feat: add tables fields 'isDisabled' and 'timeZone' to cater for more reach app request conditions
